src/requirements.txt
torch~=2.0.0
flask~=2.2.3
pillow~=9.4.0
requests~=2.28.2
salesforce-lavis
flask_api
redis